feat: use LetsEncrypt internally by default
BREAKING CHANGE:
Certificates are now an internal detail and no configuration
is required for each domain to enable SSL.

Use https://github.com/cloud-cli/le to generate certificates.

The root domain of a proxy config will be automatically mapped
to a certificate if it exists in the server running PX and LE.

The API was also simplified to shorter commands.